feat: new landing page, drop old tube list
