+++ otisweb/templates/sidebar.html
@@ -106,7 +106,7 @@
         <a title="Hanabi" class="emoji-link" href="{% url "hanabi-contests" %}">🎇</a>
       </td>
       <td>
-        <a title="Tubes" class="emoji-link" href="{% url "tube-list" %}">🧪</a>
+        <a title="OIME" class="emoji-link" href="{% url "oime-landing" %}">🧪</a>
       </td>
       <td>
         <a title="OPAL" class="emoji-link" href="{% url "opal-hunt-list" %}">🧩</a>
