+++ tubes/templates/tubes/landing.html
@@ -3,26 +3,23 @@
   OTIS Mock AIME
 {% endblock title %}
 {% block layout-content %}
-  <h2>Active Contests</h2>
-  {% if tube_list %}
-    <ul id="tube-list">
-      {% for tube in tube_list %}
-        {% if tube.accepting_signups %}
-          <li>
-            <a href="{% url "tube-join" tube.pk %}" target="_blank" rel="noopener"><b>{{ tube }}</b></a>
-            {{ tube.description|linebreaks }}
-          </li>
-        {% endif %}
-      {% endfor %}
-    </ul>
-  {% else %}
-    <div class="alert alert-warning">There is nothing to show here yet.</div>
-  {% endif %}
-  <h2>Instructions</h2>
+  <p class="fst-italic">Satisfactory. Keep cooking.</p>
   <p>
-    Click on the link above to access the problem proposals on Probase.
-    After joining, you can submit problems and participate in testsolving.
+    Welcome to the <strong>OTIS Mock AIME <s>Kitchen</s> Lab</strong>!
   </p>
+  <h2>Actions</h2>
+  <ul>
+    <li>
+      <a href="{% url "oime-proposal-list" %}">View problem bank</a>
+    </li>
+    <li>
+      <a href="{% url "oime-proposal-create" %}">Propose a problem</a>
+    </li>
+    <li>
+      <a href="{% url "oime-setup" %}">Setup</a>
+    </li>
+  </ul>
+  <h2>About the process</h2>
   <h3>Proposing</h3>
   <ul>
     <li>Use "Add Problem" once you're in the Probase collection.</li>
@@ -63,11 +60,21 @@
     time per problem when testsolving it based on the tier:
   </p>
   <ul>
-    <li>AIME 1-3: Solvers get 10 minutes</li>
-    <li>AIME 4-6: Solvers get 15 minutes</li>
-    <li>AIME 7-9: Solvers get 20 minutes</li>
-    <li>AIME 10-12: Solvers get 25 minutes</li>
-    <li>AIME 13-15: Solvers get 30 minutes</li>
+    <li>
+      <strong>Tier I:</strong> AIME 1-3, solvers get 10 minutes
+    </li>
+    <li>
+      <strong>Tier II:</strong> AIME 4-6, solvers get 15 minutes
+    </li>
+    <li>
+      <strong>Tier III:</strong> AIME 7-9, solvers get 20 minutes
+    </li>
+    <li>
+      <strong>Tier IV:</strong> AIME 10-12, solvers get 25 minutes
+    </li>
+    <li>
+      <strong>Tier V:</strong> AIME 13-15, solvers get 30 minutes
+    </li>
   </ul>
   <p>
     After you solve or give up on a proposal, you can read spoilers and comment
@@ -98,12 +105,5 @@
     because there is more time for people to upvote it
     and more opportunities for it to be testsolved.
     Thus, submitting problems well before the deadline is advised.
-  </p>
-  <h2>Thanks</h2>
-  <p>
-    Thanks to Howard Halim for writing the Probase code.
-    The source code for Probase can be found on
-    <a href="https://github.com/howard36/probase">Howard's GitHub</a>.
-    Please open bug reports or feature requests on that issue tracker.
   </p>
 {% endblock layout-content %}
